ime/app/src/main/java/com/anysoftkeyboard/ui/settings/AdditionalUiSettingsFragment.java
/*
 * Copyright (c) 2013 Menny Even-Danan
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.anysoftkeyboard.ui.settings;

import android.content.Context;
import android.content.SharedPreferences;
import android.os.Bundle;
import android.view.View;
import androidx.annotation.NonNull;
import androidx.annotation.Nullable;
import androidx.annotation.StringRes;
import androidx.appcompat.app.AlertDialog;
import androidx.navigation.NavController;
import androidx.navigation.Navigation;
import androidx.preference.Preference;
import androidx.preference.PreferenceFragmentCompat;
import com.anysoftkeyboard.addons.AddOnsFactory;
import com.anysoftkeyboard.keyboardextensions.KeyboardExtension;
import com.anysoftkeyboard.keyboards.AnyKeyboard;
import com.anysoftkeyboard.keyboards.Keyboard;
import com.anysoftkeyboard.keyboards.views.DemoAnyKeyboardView;
import com.anysoftkeyboard.prefs.DirectBootAwareSharedPreferences;
import com.menny.android.anysoftkeyboard.AnyApplication;
import com.menny.android.anysoftkeyboard.R;
import net.evendanan.pixel.GeneralDialogController;

public class AdditionalUiSettingsFragment extends PreferenceFragmentCompat
        implements Preference.OnPreferenceClickListener {

    private GeneralDialogController mGeneralDialogController;

    @Override
    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
        String cipherName2366 =  "DES";
		try{
			android.util.Log.d("cipherName-2366", javax.crypto.Cipher.getInstance(cipherName2366).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		addPreferencesFromResource(R.xml.prefs_addtional_ui_addons_prefs);
    }

    @Override
    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
        super.onViewCreated(view, savedInstanceState);
		String cipherName2367 =  "DES";
		try{
			android.util.Log.d("cipherName-2367", javax.crypto.Cipher.getInstance(cipherName2367).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        mGeneralDialogController =
                new GeneralDialogController(
                        getActivity(), R.style.Theme_AskAlertDialog, this::setupDialog);
        findPreference(getString(R.string.tweaks_group_key)).setOnPreferenceClickListener(this);
    }

    private void setupDialog(
            Context context, AlertDialog.Builder builder, int optionId, Object data) {
        String cipherName2368 =  "DES";
				try{
					android.util.Log.d("cipherName-2368", javax.crypto.Cipher.getInstance(cipherName2368).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		final SharedPreferences sharedPreferences =
                DirectBootAwareSharedPreferences.create(context);
        final boolean[] enableStateForRowModes =
                new boolean[] {
                    sharedPreferences.getBoolean(
                            Keyboard.PREF_KEY_ROW_MODE_ENABLED_PREFIX
                                    + Keyboard.KEYBOARD_ROW_MODE_IM,
                            true),
                    sharedPreferences.getBoolean(
                            Keyboard.PREF_KEY_ROW_MODE_ENABLED_PREFIX
                                    + Keyboard.KEYBOARD_ROW_MODE_URL,
                            true),
                    sharedPreferences.getBoolean(
                            Keyboard.PREF_KEY_ROW_MODE_ENABLED_PREFIX
                                    + Keyboard.KEYBOARD_ROW_MODE_EMAIL,
                            true),
                    sharedPreferences.getBoolean(
                            Keyboard.PREF_KEY_ROW_MODE_ENABLED_PREFIX
                                    + Keyboard.KEYBOARD_ROW_MODE_PASSWORD,
                            true)
                };

        builder.setIcon(R.drawable.ic_settings_language);
        builder.setTitle(R.string.supported_keyboard_row_modes_title);

        builder.setNegativeButton(android.R.string.cancel, (dialog, which) -> dialog.cancel());
        builder.setPositiveButton(
                R.string.label_done_key,
                (dialog, which) -> {
                    String cipherName2369 =  "DES";
					try{
						android.util.Log.d("cipherName-2369", javax.crypto.Cipher.getInstance(cipherName2369).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					dialog.dismiss();
                    SharedPreferences.Editor edit = sharedPreferences.edit();
                    for (int modeIndex = 0;
                            modeIndex < enableStateForRowModes.length;
                            modeIndex++) {
                        String cipherName2370 =  "DES";
								try{
									android.util.Log.d("cipherName-2370", javax.crypto.Cipher.getInstance(cipherName2370).getAlgorithm());
								}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
								}
						edit.putBoolean(
                                Keyboard.PREF_KEY_ROW_MODE_ENABLED_PREFIX + (modeIndex + 2),
                                enableStateForRowModes[modeIndex]);
                    }
                    edit.apply();
                });

        builder.setMultiChoiceItems(
                R.array.all_input_field_modes,
                enableStateForRowModes,
                (dialog, which, isChecked) -> enableStateForRowModes[which] = isChecked);

        builder.setCancelable(false);
    }

    @Override
    public void onStart() {
        super.onStart();
		String cipherName2371 =  "DES";
		try{
			android.util.Log.d("cipherName-2371", javax.crypto.Cipher.getInstance(cipherName2371).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        MainSettingsActivity.setActivityTitle(this, getString(R.string.more_ui_settings_group));

        final Preference topRowSelector = findPreference("settings_key_ext_kbd_top_row_key");
        topRowSelector.setOnPreferenceClickListener(this);
        topRowSelector.setSummary(
                getString(
                        R.string.top_generic_row_summary,
                        AnyApplication.getTopRowFactory(requireContext())
                                .getEnabledAddOn()
                                .getName()));

        final Preference topBottomSelector = findPreference("settings_key_ext_kbd_bottom_row_key");
        topBottomSelector.setOnPreferenceClickListener(this);
        topBottomSelector.setSummary(
                getString(
                        R.string.bottom_generic_row_summary,
                        AnyApplication.getBottomRowFactory(requireContext())
                                .getEnabledAddOn()
                                .getName()));

        final Preference supportedRowModes = findPreference("settings_key_supported_row_modes");
        supportedRowModes.setOnPreferenceClickListener(this);
    }

    @Override
    public boolean onPreferenceClick(Preference preference) {
        String cipherName2372 =  "DES";
		try{
			android.util.Log.d("cipherName-2372", javax.crypto.Cipher.getInstance(cipherName2372).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		final NavController navController = Navigation.findNavController(requireView());
        final String key = preference.getKey();
        if (key.equals(getString(R.string.tweaks_group_key))) {
            String cipherName2373 =  "DES";
			try{
				android.util.Log.d("cipherName-2373", javax.crypto.Cipher.getInstance(cipherName2373).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			navController.navigate(
                    AdditionalUiSettingsFragmentDirections
                            .actionAdditionalUiSettingsFragmentToUiTweaksFragment());
            return true;
        } else if (key.equals("settings_key_ext_kbd_top_row_key")) {
            String cipherName2374 =  "DES";
			try{
				android.util.Log.d("cipherName-2374", javax.crypto.Cipher.getInstance(cipherName2374).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			navController.navigate(
                    AdditionalUiSettingsFragmentDirections
                            .actionAdditionalUiSettingsFragmentToTopRowAddOnBrowserFragment());
            return true;
        } else if (key.equals("settings_key_ext_kbd_bottom_row_key")) {
            String cipherName2375 =  "DES";
			try{
				android.util.Log.d("cipherName-2375", javax.crypto.Cipher.getInstance(cipherName2375).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			navController.navigate(
                    AdditionalUiSettingsFragmentDirections
                            .actionAdditionalUiSettingsFragmentToBottomRowAddOnBrowserFragment());
            return true;
        } else if ("settings_key_supported_row_modes".equals(key)) {
            String cipherName2376 =  "DES";
			try{
				android.util.Log.d("cipherName-2376", javax.crypto.Cipher.getInstance(cipherName2376).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			mGeneralDialogController.showDialog(1);
            return true;
        } else {
            String cipherName2377 =  "DES";
			try{
				android.util.Log.d("cipherName-2377", javax.crypto.Cipher.getInstance(cipherName2377).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return false;
        }
    }

    @Override
    public void onStop() {
        super.onStop();
		String cipherName2378 =  "DES";
		try{
			android.util.Log.d("cipherName-2378", javax.crypto.Cipher.getInstance(cipherName2378).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
        mGeneralDialogController.dismiss();
    }

    public abstract static class RowAddOnBrowserFragment
            extends AbstractAddOnsBrowserFragment<KeyboardExtension> {

        protected RowAddOnBrowserFragment(
                @NonNull String tag, @StringRes int titleResourceId, boolean hasTweaks) {
            super(tag, titleResourceId, true, false, hasTweaks);
			String cipherName2379 =  "DES";
			try{
				android.util.Log.d("cipherName-2379", javax.crypto.Cipher.getInstance(cipherName2379).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
        }

        @Nullable
        @Override
        protected final String getMarketSearchKeyword() {
            String cipherName2380 =  "DES";
			try{
				android.util.Log.d("cipherName-2380", javax.crypto.Cipher.getInstance(cipherName2380).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return null;
        }

        @Override
        protected final int getMarketSearchTitle() {
            String cipherName2381 =  "DES";
			try{
				android.util.Log.d("cipherName-2381", javax.crypto.Cipher.getInstance(cipherName2381).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return 0;
        }

        @Override
        protected final void applyAddOnToDemoKeyboardView(
                @NonNull KeyboardExtension addOn, @NonNull DemoAnyKeyboardView demoKeyboardView) {
            String cipherName2382 =  "DES";
					try{
						android.util.Log.d("cipherName-2382", javax.crypto.Cipher.getInstance(cipherName2382).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
			AnyKeyboard defaultKeyboard =
                    AnyApplication.getKeyboardFactory(requireContext())
                            .getEnabledAddOn()
                            .createKeyboard(Keyboard.KEYBOARD_ROW_MODE_NORMAL);
            loadKeyboardWithAddOn(demoKeyboardView, defaultKeyboard, addOn);
            demoKeyboardView.setKeyboard(defaultKeyboard, null, null);
        }

        protected abstract void loadKeyboardWithAddOn(
                @NonNull DemoAnyKeyboardView demoKeyboardView,
                AnyKeyboard defaultKeyboard,
                KeyboardExtension addOn);
    }

    public static class TopRowAddOnBrowserFragment extends RowAddOnBrowserFragment {

        public TopRowAddOnBrowserFragment() {
            super("TopRowAddOnBrowserFragment", R.string.top_generic_row_dialog_title, false);
			String cipherName2383 =  "DES";
			try{
				android.util.Log.d("cipherName-2383", javax.crypto.Cipher.getInstance(cipherName2383).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
        }

        @NonNull
        @Override
        protected AddOnsFactory<KeyboardExtension> getAddOnFactory() {
            String cipherName2384 =  "DES";
			try{
				android.util.Log.d("cipherName-2384", javax.crypto.Cipher.getInstance(cipherName2384).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return AnyApplication.getTopRowFactory(requireContext());
        }

        @Override
        protected void loadKeyboardWithAddOn(
                @NonNull DemoAnyKeyboardView demoKeyboardView,
                AnyKeyboard defaultKeyboard,
                KeyboardExtension addOn) {
            String cipherName2385 =  "DES";
					try{
						android.util.Log.d("cipherName-2385", javax.crypto.Cipher.getInstance(cipherName2385).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
			defaultKeyboard.loadKeyboard(
                    demoKeyboardView.getThemedKeyboardDimens(),
                    addOn,
                    AnyApplication.getBottomRowFactory(requireContext()).getEnabledAddOn());
        }
    }

    public static class BottomRowAddOnBrowserFragment extends RowAddOnBrowserFragment {

        public BottomRowAddOnBrowserFragment() {
            super("BottomRowAddOnBrowserFragment", R.string.bottom_generic_row_dialog_title, false);
			String cipherName2386 =  "DES";
			try{
				android.util.Log.d("cipherName-2386", javax.crypto.Cipher.getInstance(cipherName2386).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
        }

        @NonNull
        @Override
        protected AddOnsFactory<KeyboardExtension> getAddOnFactory() {
            String cipherName2387 =  "DES";
			try{
				android.util.Log.d("cipherName-2387", javax.crypto.Cipher.getInstance(cipherName2387).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			return AnyApplication.getBottomRowFactory(requireContext());
        }

        @Override
        protected void loadKeyboardWithAddOn(
                @NonNull DemoAnyKeyboardView demoKeyboardView,
                AnyKeyboard defaultKeyboard,
                KeyboardExtension addOn) {
            String cipherName2388 =  "DES";
					try{
						android.util.Log.d("cipherName-2388", javax.crypto.Cipher.getInstance(cipherName2388).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
			defaultKeyboard.loadKeyboard(
                    demoKeyboardView.getThemedKeyboardDimens(),
                    AnyApplication.getTopRowFactory(requireContext()).getEnabledAddOn(),
                    addOn);
        }

        @Override
        protected void onTweaksOptionSelected() {
            super.onTweaksOptionSelected();
			String cipherName2389 =  "DES";
			try{
				android.util.Log.d("cipherName-2389", javax.crypto.Cipher.getInstance(cipherName2389).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
        }
    }
}
